separate aggregator rgbpp clear queue tx building.
